% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{package}
\name{katexR}
\alias{katexR}
\alias{katexR-package}
\title{KaTeX for R}
\description{
\code{katexR} is a \href{http://htmlwidgets.org}{htmlwidget} designed to make using
\href{http://khan.github.io/KaTeX/}{KaTeX} easy in R.  Since \code{katexR}
is an \code{htmlwidget}, we can use \code{KaTeX} in nearly all R contexts:
console, RStudio Viewer, browser, and Shiny.
}
\details{
\strong{Please use the function \code{\link{katex}}}.  The very early version used the harder
to type and now  \code{katexR}.
}

