% Generated by roxygen2 (4.0.2): do not edit by hand
\name{katexR}
\alias{katexR}
\title{KaTeX for R}
\usage{
katexR(katex = "", tag = "div", style = NULL)
}
\description{
\code{katexR} is a \href{http://htmlwidgets.org}{htmlwidget} designed to make using
\href{http://khan.github.io/KaTeX/}{KaTeX} easy in R.  Since \code{katexR}
is an \code{htmlwidget}, we can use \code{KaTeX} in nearly all R contexts -
console, RStudio Viewer, browser, and Shiny.
}

